refactor: move in-memory repo implementation into appropriate file - move domain validations to appropriate layer
